+++ modules/rest/src/main/java/org/entrystore/rest/resources/AbstractMetadataResource.java
@@ -162,13 +162,14 @@ public abstract class AbstractMetadataResource extends BaseResource {
 						result.setModificationDate(travResult.getLatestModified());
 					}
 				} else {
-					if (getMetadata() == null) {
+					// MergedMetadataResource does not implement getMetadata()
+					if (getMetadata() == null && getMetadataGraph() == null) {
 						getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND);
 						return null;
 					}
 
 					if (graphQuery != null) {
-						Model graphQueryResult = applyGraphQuery(graphQuery, getMetadata().getGraph());
+						Model graphQueryResult = applyGraphQuery(graphQuery, getMetadataGraph());
 						if (graphQueryResult != null) {
 							result = getRepresentation(graphQueryResult, prefFormat);
 						} else {
@@ -176,7 +177,7 @@ public abstract class AbstractMetadataResource extends BaseResource {
 							return null;
 						}
 					} else {
-						result = getRepresentation(getMetadata().getGraph(), prefFormat);
+						result = getRepresentation(getMetadataGraph(), prefFormat);
 					}
 				}
 
@@ -225,7 +226,7 @@ public abstract class AbstractMetadataResource extends BaseResource {
 			}
 
 			if (getMetadata() == null) {
-				getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND);
+				getResponse().setStatus(Status.CLIENT_ERROR_METHOD_NOT_ALLOWED);
 				return;
 			}
 
@@ -247,7 +248,7 @@ public abstract class AbstractMetadataResource extends BaseResource {
 			}
 
 			if (getMetadata() == null) {
-				getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND);
+				getResponse().setStatus(Status.CLIENT_ERROR_METHOD_NOT_ALLOWED);
 				return;
 			}
 
@@ -299,9 +300,19 @@ public abstract class AbstractMetadataResource extends BaseResource {
 	}
 
 	/**
-	 * @return Returns the relevant metadata graph. May be null.
+	 * @return Returns the metadata object. May be null.
 	 */
 	protected abstract Metadata getMetadata();
+
+	/**
+	 * @return Returns the metadata graph. May be null.
+	 */
+	protected Model getMetadataGraph() {
+		if (getMetadata() != null) {
+			return getMetadata().getGraph();
+		}
+		return null;
+	}
 
 	/**
 	 * @return Returns the modification date of the metadata graph. Only external metadata has its own date;
@@ -418,15 +429,13 @@ public abstract class AbstractMetadataResource extends BaseResource {
 			rc = sr.getConnection();
 			rc.add(graph);
 			GraphQuery gq = rc.prepareGraphQuery(QueryLanguage.SPARQL, query);
-			gq.setMaxQueryTime(10); // 10 seconds, TODO: make this configurable
+			gq.setMaxExecutionTime(10); // 10 seconds, TODO: make this configurable
 			result = Iterations.addAll(gq.evaluate(), new LinkedHashModel());
 			log.info("Graph query took " + (new Date().getTime() - before.getTime()) + " ms");
-		} catch (RepositoryException e) {
+		} catch (RepositoryException | QueryEvaluationException e) {
 			log.error(e.getMessage());
 		} catch (MalformedQueryException mfqe) {
 			log.debug(mfqe.getMessage());
-		} catch (QueryEvaluationException qee) {
-			log.error(qee.getMessage());
 		} finally {
 			if (rc != null) {
 				try {
@@ -444,10 +453,10 @@ public abstract class AbstractMetadataResource extends BaseResource {
 		return result;
 	}
 
-	private static RDFFormat RDFJSON_WITH_APPLICATION_JSON
+	private static final RDFFormat RDFJSON_WITH_APPLICATION_JSON
 		= new RDFFormat("RDF/JSON", List.of("application/json"), StandardCharsets.UTF_8, List.of("json"), SimpleValueFactory.getInstance().createIRI("http://www.w3.org/ns/formats/RDF_JSON"), false, true, false);
 
-	private String getFileExtensionForMediaType(MediaType mt) {
+	protected static String getFileExtensionForMediaType(MediaType mt) {
 		Optional<RDFFormat> rdfFormat = RDFFormat.matchMIMEType(mt.getName(), Arrays.asList(
 				RDFFormat.RDFXML,
 				RDFFormat.NTRIPLES,
@@ -467,4 +476,5 @@ public abstract class AbstractMetadataResource extends BaseResource {
 		}
 		return "rdf";
 	}
+
 }